Make BoundingBoxDialog the example of how to build one, remove BoundingBoxMamutMain
